requirements.txt
# pip install -r requirements.txt
numpy==1.21.0
matplotlib==3.4.3
fastapi==0.68.0
uvicorn==0.15.0
python-dotenv==0.19.1
psycopg2-binary==2.9.1
Faker==8.12.1
google-cloud==0.34.0
google-cloud-storage==1.42.3
google-cloud-bigquery==2.30.1
azure==5.0.0
azure-storage-blob==12.9.0
azure-cosmos==4.2.0
azure-identity==1.7.1
boto3==1.20.23
awscli==1.22.10
